Markings Bundle From Impstation (#386)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

markings from ImpStation

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This is an enormous PR bundling nearly every marking and hairstyle from
Impstation.

Port of a port.
Work initially done for Frontier Station by HiveHum
https://github.com/new-frontiers-14/frontier-station-14/pull/2771

Minor resprite to slime people that alters their translucency and
removes their mouth and nose. Nose and mouth for slimes are now
markings.

List of ImpStation ports by various contributors:
https://github.com/impstation/imp-station-14/pull/223
https://github.com/impstation/imp-station-14/pull/413
https://github.com/impstation/imp-station-14/pull/548
https://github.com/impstation/imp-station-14/pull/550
https://github.com/impstation/imp-station-14/pull/552
https://github.com/impstation/imp-station-14/pull/563
https://github.com/impstation/imp-station-14/pull/591
https://github.com/impstation/imp-station-14/pull/765
https://github.com/impstation/imp-station-14/pull/833
https://github.com/impstation/imp-station-14/pull/842
https://github.com/impstation/imp-station-14/pull/906
https://github.com/impstation/imp-station-14/pull/951
https://github.com/impstation/imp-station-14/pull/1090
https://github.com/impstation/imp-station-14/pull/1096
https://github.com/impstation/imp-station-14/pull/1181
https://github.com/impstation/imp-station-14/pull/1204
https://github.com/impstation/imp-station-14/pull/1234
https://github.com/impstation/imp-station-14/pull/1240
https://github.com/impstation/imp-station-14/pull/1243
https://github.com/impstation/imp-station-14/pull/1248

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Impstation
- add: From the IMP sector, over a hundred new hair styles and markings!
- tweak:Vox and Slimes have received a minor visual refresh.

---------

Co-authored-by: hivehum <ketchupfaced@gmail.com>
Co-authored-by: Whatstone <whatston3@gmail.com>