updated gas usages to no longer typecast in yml
Done prior, and were functional. Possibly to do with inheritance or changing the VV variable.
Changed to no longer end in f(declaring them as floats rather than doubles) and instead allowing the component to implicitly cast them as such.
No errors (save on cannon projectile, which is long standing.)