The Denu: Custom Hearing Range (#1385)
Added the ability to set a custom hearing range for local, emotes, and
LOOC in the Denu!
Thank you to Cami for the visuals and UI code <3

<img width="1919" height="1073" alt="image"
src="https://github.com/user-attachments/assets/7e852c7d-444c-4f5b-95a2-4c20688bce52"
/>

**Changelog**
<!-- Add a Changelog entry to make players aware of new features or
changes that could affect gameplay.
Make sure to read the guidelines and take this Changelog template out of
the comment block in order for it to show up.
Changelog must have a :cl: symbol, so the bot recognizes the changes and
adds them to the game's changelog. -->
<!--
:cl:
- add: Added fun!
- remove: Removed fun!
- tweak: Changed fun!
- fix: Fixed fun!
-->
:cl: sleepyyapril, Cami
- add: Added the option to set a custom hearing range in the Denu! No
more hearing 2000 people surrounding you!

---------

Co-authored-by: Cam <Nop>