Sleeping system invocations commented out of IPC death system
Not only does this mean we don't have an SSD Indicator issues thrown when the IPC is marked as SSD and sleep is applied, but now when a player gets EMP'd they are not helplessly at the whim of random chance.
They can still speak and use their hands, just not move.
The One-Hit KO of IPCs is over, rejoice.
Also EOF newline added to categories.